requirements.txt
numpy
hnswlib
Pillow
transformers
torch
sentencepiece
accelerate